Move multiple components to more logical locations
