+++ app/src/components/feature/rightpanel/comments/comment/Comment.scss
