+++ app/src/components/feature/rightpanel/historyflow/HistoryFlow.component.jsx
@@ -20,13 +20,13 @@ import SearchEdge from 'components/feature/advancedsearch/searchedge/SearchEdge.
 import 'overlayscrollbars/styles/overlayscrollbars.css';
 import ReactFlow, { MiniMap } from 'react-flow-renderer';
 import AutoSizer from 'react-virtualized-auto-sizer';
-import DatasetNode from '../advancedsearch/datasetNode/Dataset.component';
-import SchemaEdge from '../advancedsearch/searchedge/SearchEdge.component';
-import SearchNode from '../advancedsearch/searchnode/SearchNode.component';
-import HistoryNode from '../historyNode/HistoryNode.component';
-import OverviewCustomEdge from '../overviewschemaedge/OverviewSchemaEdge.component';
-import OverviewSchemaNode from '../overviewschemanode/OverviewSchemaNode.component';
-import SchemaNode from '../schemanode/SchemaNode.component';
+import DatasetNode from '../../advancedsearch/datasetNode/Dataset.component';
+import SchemaEdge from '../../advancedsearch/searchedge/SearchEdge.component';
+import SearchNode from '../../advancedsearch/searchnode/SearchNode.component';
+import HistoryNode from '../../historyNode/HistoryNode.component';
+import OverviewCustomEdge from '../../overviewschemaedge/OverviewSchemaEdge.component';
+import OverviewSchemaNode from '../../overviewschemanode/OverviewSchemaNode.component';
+import SchemaNode from '../../schemanode/SchemaNode.component';
 
 export function HistoryFlow() {
     const store = useContext(RootStoreContext);
