+++ app/src/components/feature/rightpanel/table/globalFilter/GlobalFilter.component.jsx
