+++ app/src/components/feature/rightpanel/table/table.scss
